バグ修正

+++ sp_index.rb
@@ -306,18 +306,20 @@ end
     decide_e_schedule(ed)
     e_num=@num_i
     #printf("test: sd%s, ed%s, s_num%s e_num%s\n", sd, ed, s_num, e_num)
-    for i in s_num..e_num
-      if @title[i]=="sleep"
-        #printf("test: sleep is %s\n", @s_day[i])
-        db = SQLite3::Database.new('scheduler.db')
-          db.execute('delete from schedule where id=?', @id[i])
-        db.close
+    if s_num.to_i>1 && e_num.to_i>1
+      for i in s_num..e_num
+        if @title[i]=="sleep"
+          #printf("test: sleep is %s\n", @s_day[i])
+          db = SQLite3::Database.new('scheduler.db')
+            db.execute('delete from schedule where id=?', @id[i])
+          db.close
+        end
       end
     end
   end
 
   def sleep_t
-    day = prevday(@today)
+    day = @today
     #printf("test: def sleep_t, dat=%s\n",day)
     db = SQLite3::Database.new('scheduler.db')
     db.results_as_hash = true
@@ -335,7 +337,11 @@ end
     null_sleep(sd, ed)
     for i in 0..@inputdays.to_i - 1
       s_day = day
-      e_day = nextday(day)
+      if st.to_i>et.to_i
+        e_day = nextday(day)
+      else
+        e_day=day
+      end
       if search_same('sleep', s_day, st, e_day, et) == 0
         db.execute('insert into schedule  (title, s_day, s_time, e_day, e_time, st) values(?, ?, ?, ?, ?, ?)', 'sleep', s_day, st, e_day, et, 's')
       end
@@ -362,267 +368,6 @@ end
     #ログ情報をtable"log"にいれる
     wday=["sun", "mon", "tue","wed", "the", "fri", "sat", "sun"]
     d.wday
-  end
-
-  def add_db_task(i, inputday, st, et)
-  	db = SQLite3::Database.new('scheduler.db')
-  	db.execute('insert into schedule  (title, s_day, s_time, e_day, e_time, category, st, completed) values(?, ?, ?, ?, ?, ?, ?, ?)', @t_title[i], inputday, st, inputday, et, @t_category[i], @t_id[i], '0')
-    if @l_tasktime[i]=="00:00"
-      @l_tasktime[i]=to_h(to_min(et).to_i-to_min(st).to_i)
-    else
-      @l_tasktime[i]=to_h(to_min(@l_tasktime[i]).to_i + (to_min(et).to_i-to_min(st).to_i))
-    end
-  	db.execute('update task set located = ? where id=?', @l_tasktime[i], @t_id[i])
-  	db.close
-    read_task
-  	check_tasktime(@t_id[i])
-  end
-
-
-  def task_add_time(s_time, e_time, b_time, task, c, inputday, i, flag)
-    b_time = b_time.to_i-20
-    if b_time.to_i<10
-    elsif b_time.to_i>0
-      #stとetを決める sの追加はput_taskでやったもらったほうがよいかも
-      if flag=="0"
-        if @c_min[c].to_i>b_time.to_i
-        else
-          if @c_min[c].to_i>task
-            task=@c_min[c].to_i
-            add_time=@c_min[c].to_i-task.to_i
-            new_tasktime=to_h(to_min(@tasktime[i]).to_i+add_time.to_i)
-            @tasktime[i]=new_tasktime
-            db = SQLite3::Database.new('scheduler.db')
-              db.execute('update task set t_time = ? where id=?', @tasktime[i], @t_id[i])
-            db.close
-            task=@c_min[c].to_i
-            #printf("test: L403 new_task:%s\n",task)
-              #ここまで、タスクがc_min以下だった場合タスク時間をc_minの差分分増やす
-          end
-        if b_time.to_i >task.to_i
-          #printf("test: call392 c_max is %s\n",@c_max[c])
-          if task.to_i > @c_max[c].to_i
-          #printf("test: call394n")
-            #ad c_max
-            st=to_h(to_min(e_time).to_i+10)
-            et=to_h(to_min(st).to_i+@c_max[c].to_i)
-            add_db_task(i, inputday, st, et)
-          elsif task.to_i <= @c_max[c].to_i
-            #ad task
-            st=to_h(to_min(e_time).to_i+10)
-            et=to_h(to_min(st).to_i+task.to_i)
-            #printf("test: call403\n")
-            add_db_task(i, inputday, st, et)
-          end
-        else
-          if b_time.to_i>@c_max[c].to_i
-            #ad c_mac
-            st=to_h(to_min(e_time).to_i+10)
-            et=to_h(to_min(st).to_i+@c_max[c].to_i)
-            #printf("test: call411\n")
-            add_db_task(i, inputday, st, et)
-          elsif b_time.to_i < @c_max[c].to_i
-            #ad b_time
-            st=to_h(to_min(e_time).to_i+10)
-            et=to_h(to_min(st).to_i+b_time.to_i)
-            #printf("test: call417(st:%s, et%s, b_time:%s)\n", st, et, b_time)
-            add_db_task(i, inputday, st, et)
-          end
-        end
-      end
-    else
-      #flag==1のとき
-      if@c_max[c].to_i>task.to_i
-        if b_time.to_i+60>task.to_i
-          st=to_h(to_min(e_time).to_i+30)
-          et=to_h(to_min(st).to_i+task.to_i)
-          add_db_task(i, inputday, st, et)
-        elsif b_time.to_i+60<=task.to_i
-        end
-      end
-    end
-  end
-  end
-
-
-
-  def put_task
-    read_schedule
-    read_task
-    decide_s_schedule(@today)
-    s=@num_i.to_i
-  #  printf("call put_task\n")
-    #タスクのeventの追加処理
-    #スケジュールを日付順に並べ替え、@num_i番目が今日＋１日目のスケジュール
-    #@s_day[s]が今日＋１日めのスケジュール
-    endday=decide_eday
-    day=decide_sday
-    #s_day[@num_i]~enddayまでの間にスケジュールを追加する
-    #タスクiのカテゴリは@c_name[c]である
-    read_task
-    read_category
-    i=0
-  #  printf("t_num:%s\n",@t_num)
-    until i==@t_num
-      check_tasktime(@t_id[i])
-      #printf("test:L472 iは%s,　resttimeは%s\n", i, $resttime)
-      #タスクの残り作業時刻の計測
-      #$resttimeが変数
-      #printf("%s i:%s, 残作業時刻は%s\n", @t_title[i], i, $resttime)
-      if $resttime!=0
-        endday=@te_day[i]
-        for j in 0.. @c_num.to_i-1
-          #カテゴリテーブルのカテゴリj
-          if @t_category[i]==@c_name[j]
-            #現在のタスクiのカテゴリ名がカテゴリとヒットした時
-            c=j
-            break;
-          end
-        end
-        #↑カテゴリ検索end
-        checkday=@today
-        #printf("test:checkday=%s\n", checkday)
-        if prevday(checkday)==endday && $resttime>0
-          #開いた当日が締切日の場合
-          #printf("L490 開いた当日が締切日の場合\n")
-          for j in 0..@num.to_i-1
-            if endday==@e_day[j]
-              s=j
-              break
-            end
-          end
-          if @e_day[s]==@s_day[s+1]
-            b_time=to_min(@s_time[s+1]).to_i-to_min(@e_time[s]).to_i
-            task_add_time(@s_time[s+1], @e_time[s], b_time,$resttime, c, endday, i, "0")
-          end
-        end
-        #当日締切end
-        until chint(checkday)>chint(endday)
-          #printf("(%s:%s) test:checkyday=%s, %s:@e_day[s]=%s\n", i, @t_title[i],  checkday, @title[s],@e_day[s])
-          if $resttime<=0
-            #printf("test: break %s\n", i)
-            break;
-          end
-          #  printf("checkday : %s, endday : %s\n",checkday, endday)
-          while chint(checkday).to_i>chint(@e_day[s]).to_i
-            s=s+1
-          end
-          #printf("test: %s, %s\n",@e_day[s], @s_day[s+1])
-          if checkday==@e_day[s]
-            #指定日に予定がある
-            if @e_day[s]==@s_day[s+1]
-              #その次の予定も同日である
-              if @category[s]==@c_name[c]
-                b_time=to_min(@s_time[s+1]).to_i-to_min(@e_time[s]).to_i
-                #printf("519 / test: b_time is %s\n", b_time)
-                task_add_time(@s_time[s+1], @e_time[s], b_time,$resttime, c, checkday, i, "0")
-                s=s+1
-              elsif @category[s]==nil
-                b_time=to_min(@s_time[s+1]).to_i-to_min(@e_time[s]).to_i
-                #printf("524 / b_time is %s\n", b_time)
-                task_add_time(@s_time[s+1], @e_time[s], b_time, $resttime, c, checkday, i, "0")
-                s=s+1
-              else
-                s=s+1
-              end
-            else
-              #printf("test: L518//%s\n", @e_time[s])
-              if to_min(@e_time[s]).to_i < to_min("12:00")
-                s=s+1#タスクを配置する
-              end
-            end
-          else
-            #指定日に予定がない
-          end
-          if chint(checkday).to_i < chint(@e_day[s]).to_i
-            checkday=nextday(checkday)
-          end
-        end
-        #checkdayが締め切りになるまで
-        if $resttime!=0
-          #printf("test 530/call not_same_category:%s\n",i)
-          #以下、カテゴリどうし近しいものが無かった場合強制的に追加
-          checkday=@today
-          decide_s_schedule(@today)
-          s=@num_i.to_i
-            #printf("test: L536/ endday=%s\n", endday)
-          until chint(checkday)>chint(endday)
-            #printf("test: L536/ checkday=%s\n", checkday)
-            if $resttime==0
-              break;
-            end
-            while chint(checkday).to_i>chint(@e_day[s]).to_i
-              s=s+1
-            end
-            if chint(checkday).to_i < chint(@e_day[s]).to_i
-              checkday=nextday(checkday)
-            end
-            if checkday==@e_day[s]
-              if @e_day[s]==@s_day[s+1]
-                b_time=to_min(@s_time[s+1]).to_i-to_min(@e_time[s]).to_i
-                #printf("565 / b_time is %s\n", b_time)
-                task_add_time(@s_time[s+1], @e_time[s], b_time, $resttime, c, checkday, i, s)
-                s=s+1
-              else
-                #同日に次の予定がない場合
-              end
-            else
-              #指定日に予定がない
-            end
-          end
-        end
-        i=i+1
-        decide_s_schedule(@today)
-        s=@num_i.to_i
-      else
-        #resttimeが0の時
-        i=i+1
-        decide_s_schedule(@today)
-        s=@num_i.to_i
-      end
-      #↑resttime!=0 end
-    end
-    #↑until i==@t_num end
-  end
-
-  def null_task
-    #printf("test:[call null_task]!\n")
-    read_task
-    searchday=@today
-    decide_s_schedule(@today)
-    s=0
-    db = SQLite3::Database.new('scheduler.db')
-      #printf("test:@s_name[%s]:%s\n", s, @title[s])
-      #printf("test: @num-1=%s\n",@num-1)
-    while s!=@num-1
-      #printf("test:@s_name[%s]:%s\n", s, @title[s])
-      if chint(searchday).to_i < chint(@e_day[s]).to_i
-        searchday=nextday(searchday)
-      end
-      #printf("test: s=%s, com=%s searchday:%s, sday[s]:%s\n",s, @com[s], searchday, @s_day[s])
-      #printf("test: prevday=%s\n", prevday(@today))
-      if prevday(@today)!=@s_day[s] && @com[s]==0
-        #printf("test: delete %s(id:%s)\n",@title[s], @id[s])
-        db.execute('delete from schedule where id=?', @id[s])
-        del_min=to_min(@e_time[s]).to_i-to_min(@s_time[s]).to_i
-        for i in 0.. @t_num-1
-          if @st[s]==@t_id[i]
-            id=i
-            break
-          end
-        end
-        #printf("test:%s del_min=%s, id=%s\n", @t_title[id], del_min, @t_id[id])
-        #printf("id:%s, \n", id)
-        new_located=to_h(to_min(@l_tasktime[id]).to_i-del_min.to_i)
-        #printf("test:%s located=%s, id=%s\n", @t_title[id], new_located, @t_id[id])
-        db.execute('update task set located = ? where id=?', new_located, @t_id[id])
-        @l_tasktime[id]=new_located
-        #printf("test: @l_tasktime[%s]:%s\n",id, @l_tasktime[id])
-      end
-      s=s+1
-    end
-    db.close
-    put_task
   end
 
   def view_event
@@ -691,7 +436,6 @@ for i in 0..2
 #  event.eating_t(eat_st[i], eat_et[i])
 end
 # event.overlap_event("2015-11-04", "2015-11-04", "15:00", "17:00")
-event.null_task
 
 event.view_event
 print_t('sp_js2.txt')
